refactor(api): split praise utils into 'core' and 'score' modules
